# docs/openapi.yaml
basePath: /api
...
host: localhost:8090
info:
  contact: {}
  description: Self-hosted CTF platform API. Most write endpoints require authentication
    via JWT cookie (auth_token).
  title: hCTF2 API
  version: 1.0.0
paths:
  /admin/categories:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Category name
        in: formData
        name: name
        required: true
        type: string
      - description: Sort order (default 0)
        in: formData
        name: sort_order
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML fragment of the new category row
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create a new challenge category (admin only)
      tags:
      - Admin
  /admin/categories/{id}:
    delete:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Empty response on success
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Delete a challenge category (admin only)
      tags:
      - Admin
    put:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category name
        in: formData
        name: name
        required: true
        type: string
      - description: Sort order
        in: formData
        name: sort_order
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML fragment of the updated category row
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Update a challenge category (admin only)
      tags:
      - Admin
...
  /admin/custom-code:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              body_end_html:
                type: string
              head_html:
                type: string
              motd:
                type: string
              pages_json:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get current custom HTML/JS injection settings (admin only)
      tags:
      - Admin
    put:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: HTML to inject in <head>
        in: formData
        name: head_html
        type: string
      - description: HTML to inject before </body>
        in: formData
        name: body_end_html
        type: string
      - description: JSON config for per-page injection
        in: formData
        name: pages_json
        type: string
      - description: Message of the day
        in: formData
        name: motd
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Settings updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Update custom HTML/JS injection settings (admin only)
      tags:
      - Admin
  /admin/difficulties:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Difficulty name
        in: formData
        name: name
        required: true
        type: string
      - description: Badge color Tailwind classes
        in: formData
        name: color
        type: string
      - description: Text color Tailwind class
        in: formData
        name: text_color
        type: string
      - description: Sort order (default 0)
        in: formData
        name: sort_order
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML fragment of the new difficulty row
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create a new difficulty level (admin only)
      tags:
      - Admin
  /admin/difficulties/{id}:
    delete:
      parameters:
      - description: Difficulty ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Empty response on success
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Delete a difficulty level (admin only)
      tags:
      - Admin
    put:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Difficulty ID
        in: path
        name: id
        required: true
        type: string
      - description: Difficulty name
        in: formData
        name: name
        required: true
        type: string
      - description: Badge color Tailwind classes
        in: formData
        name: color
        type: string
      - description: Text color Tailwind class
        in: formData
        name: text_color
        type: string
      - description: Sort order
        in: formData
        name: sort_order
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML fragment of the updated difficulty row
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Update a difficulty level (admin only)
      tags:
      - Admin
...
  /admin/users:
    get:
      produces:
      - text/html
      responses:
        "200":
          description: HTML fragments with user rows
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: List all users (admin only)
      tags:
      - Admin
  /admin/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Empty response on success
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Delete a user account (admin only)
      tags:
      - Admin
  /admin/users/{id}/admin:
    put:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML fragment of the updated user row
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Toggle admin status for a user (admin only)
      tags:
      - Admin
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"